+++ tests/testthat/test_survival_surv_parametric.R
@@ -62,7 +62,7 @@ test_that("missing", {
   expect_error(learner$predict(tsk("lung")))
 })
 
-test_that("quantile", {
+test_that("quantile type", {
   learner = lrn("surv.parametric", dist = "weibull", type = "aft")$train(task)
   p = lrn("surv.parametric", dist = "weibull", type = "aft")$train(task)$predict(task)
   quantile = p$distr$quantile(c(0.2, 0.8))
@@ -78,4 +78,68 @@ test_that("quantile", {
   p = lrn("surv.parametric", dist = "weibull", type = "po")$train(task)$predict(task)
   quantile = p$distr$quantile(0.5)
   expect_equal(unlist(p$distr$cdf(quantile), use.names = FALSE), rep(0.5, 227))
+})
+
+test_that("quantile dist", {
+  learner = lrn("surv.parametric", dist = "weibull", type = "aft")$train(task)
+  p = learner$predict(task)
+  quantile = p$distr$quantile(c(0.2, 0.8))
+  expect_equal(matrix(t(quantile), ncol = 2),
+               predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8)))
+
+  learner = lrn("surv.parametric", dist = "exponential", type = "aft")$train(task)
+  p = learner$predict(task)
+  quantile = p$distr$quantile(c(0.2, 0.8))
+  expect_equal(matrix(t(quantile), ncol = 2),
+               predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8)))
+
+  learner = lrn("surv.parametric", dist = "gaussian", type = "tobit")$train(task)
+  p = learner$predict(task)
+  quantile = p$distr$quantile(c(0.2, 0.8))
+  expect_equal(matrix(t(quantile), ncol = 2),
+               predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8)))
+
+  learner = lrn("surv.parametric", dist = "lognormal")$train(task)
+  p = learner$predict(task)
+  quantile = p$distr$quantile(c(0.2, 0.8))
+  expect_equal(matrix(t(quantile), ncol = 2),
+               predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8)))
+
+  learner = lrn("surv.parametric", dist = "loglogistic")$train(task)
+  p = learner$predict(task)
+  quantile = p$distr$quantile(c(0.2, 0.8))
+  expect_equal(matrix(t(quantile), ncol = 2),
+               predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8)))
+})
+
+test_that("cdf dist", {
+  learner = lrn("surv.parametric", dist = "weibull", type = "aft")$train(task)
+  p = learner$predict(task, row_ids = 151:200)
+  cdf = predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8))[151:200,]
+  expect_equal(unname(as.matrix(p$distr$cdf(data = t(cdf)))),
+               matrix(c(rep(0.2, 50), rep(0.8, 50)), byrow = TRUE, nrow = 2))
+
+  learner = lrn("surv.parametric", dist = "exponential", type = "aft")$train(task)
+  p = learner$predict(task, row_ids = 151:200)
+  cdf = predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8))[151:200,]
+  expect_equal(unname(as.matrix(p$distr$cdf(data = t(cdf)))),
+               matrix(c(rep(0.2, 50), rep(0.8, 50)), byrow = TRUE, nrow = 2))
+
+  learner = lrn("surv.parametric", dist = "gaussian", type = "tobit")$train(task)
+  p = learner$predict(task, row_ids = 151:200)
+  cdf = predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8))[151:200,]
+  expect_equal(unname(as.matrix(p$distr$cdf(data = t(cdf)))),
+               matrix(c(rep(0.2, 50), rep(0.8, 50)), byrow = TRUE, nrow = 2))
+
+  learner = lrn("surv.parametric", dist = "lognormal")$train(task)
+  p = learner$predict(task, row_ids = 151:200)
+  cdf = predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8))[151:200,]
+  expect_equal(unname(as.matrix(p$distr$cdf(data = t(cdf)))),
+               matrix(c(rep(0.2, 50), rep(0.8, 50)), byrow = TRUE, nrow = 2))
+
+  learner = lrn("surv.parametric", dist = "loglogistic")$train(task)
+  p = learner$predict(task, row_ids = 151:200)
+  cdf = predict(learner$model$fit, type = "quantile", p = c(0.2, 0.8))[151:200,]
+  expect_equal(unname(as.matrix(p$distr$cdf(data = t(cdf)))),
+               matrix(c(rep(0.2, 50), rep(0.8, 50)), byrow = TRUE, nrow = 2))
 })